Struct amplify_num::u256

pub struct u256(_);

Large integer type

The type is composed of little-endian ordered 64-bit words, which represents its inner representation.

Implementations

impl u256

pub fn as_ptr(&self) -> *const u64

Converts the object to a raw pointer

pub fn as_mut_ptr(&mut self) -> *mut u64

Converts the object to a mutable raw pointer

pub fn as_inner(&self) -> &[u64; 4]

Returns the underlying array of words constituting large integer

pub fn into_inner(self) -> [u64; 4]

Returns the underlying array of words constituting large integer

pub const fn from_inner(array: [u64; 4]) -> Self

Constructs integer type from the underlying array of words.

impl u256

pub const ZERO: u256 = _

Zero value

pub const ONE: u256 = _

Value for 1

pub const BITS: u32 = 256u32

Bit dimension

pub const BYTES: u8 = 32u8

Length of the integer in bytes

pub const INNER_LEN: u8 = 4u8

Length of the inner representation in 64-bit words

pub fn bit(&self, index: usize) -> bool

Returns whether specific bit number is set to 1 or not

pub fn low_u32(&self) -> u32

Returns lower 32 bits of the number as u32

pub fn low_u64(&self) -> u64

Returns lower 64 bits of the number as u32

pub fn leading_ones(&self) -> u32

Returns the number of leading ones in the binary representation of self.

pub fn leading_zeros(&self) -> u32

Returns the number of leading zeros in the binary representation of self.

pub fn trailing_ones(&self) -> u32

Returns the number of trailing ones in the binary representation of self.

pub fn trailing_zeros(&self) -> u32

Returns the number of trailing zeros in the binary representation of self.

pub fn is_zero(&self) -> bool

pub fn abs(self) -> u256

pub fn from_be_bytes(bytes: [u8; 32]) -> u256

Creates the integer value from a byte array using big-endian encoding

pub fn from_be_slice(bytes: &[u8]) -> Result<u256, ParseLengthError>

Creates the integer value from a byte slice using big-endian encoding

pub fn from_le_bytes(bytes: [u8; 32]) -> u256

Creates the integer value from a byte array using little-endian encoding

pub fn from_le_slice(bytes: &[u8]) -> Result<u256, ParseLengthError>

Creates the integer value from a byte slice using little-endian encoding

pub fn to_be_bytes(self) -> [u8; 32]

Convert the integer into a byte array using big-endian encoding

pub fn to_le_bytes(self) -> [u8; 32]

Convert a integer into a byte array using little-endian encoding

impl u256

pub fn checked_add<T>(self, other: T) -> Option<u256>where
T: Into<u256>,

Checked integer addition. Computes self + rhs, returning None if overflow occurred.

pub fn saturating_add<T>(self, other: T) -> u256where
T: Into<u256>,

Saturating integer addition. Computes self + rhs, saturating at the numeric bounds instead of overflowing.

pub fn overflowing_add<T>(self, other: T) -> (u256, bool)where
T: Into<u256>,

Calculates self + rhs

Returns a tuple of the addition along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub fn wrapping_add<T>(self, other: T) -> u256where
T: Into<u256>,

Wrapping (modular) addition. Computes self + rhs, wrapping around at the boundary of the type.

pub fn checked_sub<T>(self, other: T) -> Option<u256>where
T: Into<u256>,

Checked integer subtraction. Computes self - rhs, returning None if overflow occurred.

pub fn saturating_sub<T>(self, other: T) -> u256where
T: Into<u256>,

Saturating integer subtraction. Computes self - rhs, saturating at the numeric bounds instead of overflowing.

pub fn overflowing_sub<T>(self, other: T) -> (u256, bool)where
T: Into<u256>,

Calculates self - rhs

Returns a tuple of the subtraction along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub fn wrapping_sub<T>(self, other: T) -> u256where
T: Into<u256>,

Wrapping (modular) subtraction. Computes self - rhs, wrapping around at the boundary of the type.

pub fn checked_mul<T>(self, other: T) -> Option<u256>where
T: Into<u256>,

Checked integer multiplication. Computes self * rhs, returning None if overflow occurred.

pub fn saturating_mul<T>(self, other: T) -> u256where
T: Into<u256>,

Saturating integer multiplication. Computes self * rhs, saturating at the numeric bounds instead of overflowing.

pub fn wrapping_mul<T>(self, other: T) -> u256where
T: Into<u256>,

Wrapping (modular) multiplication. Computes self * rhs, wrapping around at the boundary of the type.

pub fn overflowing_div<T>(self, other: T) -> (u256, bool)where
T: Into<u256>,

Calculates self / rhs

Returns a tuple of the divisor along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub fn wrapping_div<T>(self, other: T) -> u256where
T: Into<u256>,

Wrapping (modular) division. Calculates self / rhs, wrapping around at the boundary of the type.

The only case where such wrapping can occur is when one divides MIN / -1 on a signed type (where MIN is the negative minimal value for the type); this is equivalent to -MIN, a positive value that is too large to represent in the type. In such a case, this function returns MIN itself.

pub fn checked_div<T>(self, other: T) -> Option<u256>where
T: Into<u256>,

Checked integer division. Computes self / rhs, returning None if rhs == 0 or the division results in overflow.

pub fn saturating_div<T>(self, other: T) -> u256where
T: Into<u256>,

Saturating integer division. Computes self / rhs, saturating at the numeric bounds instead of overflowing.

pub fn overflowing_rem<T>(self, other: T) -> (u256, bool)where
T: Into<u256>,

Calculates the remainder when self is divided by rhs.

Returns a tuple of the remainder after dividing along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then 0 is returned.

pub fn wrapping_rem<T>(self, other: T) -> u256where
T: Into<u256>,

Wrapping (modular) remainder. Computes self % rhs, wrapping around at the boundary of the type.

Such wrap-around never actually occurs mathematically; implementation artifacts make x % y invalid for MIN / -1 on a signed type (where MIN is the negative minimal value). In such a case, this function returns 0.

pub fn checked_rem<T>(self, other: T) -> Option<u256>where
T: Into<u256>,

Checked integer remainder. Computes self % rhs, returning None if rhs == 0 or the division results in overflow.

pub fn div_euclid<T>(self, other: T) -> u256where
T: Into<u256>,

Calculates the quotient of Euclidean division of self by rhs.

This computes the integer q such that self = q * rhs + r, with r = self.rem_euclid(rhs) and 0 <= r < abs(rhs).

In other words, the result is self / rhs rounded to the integer q such that self >= q * rhs. If self > 0, this is equal to round towards zero (the default in Rust); if self < 0, this is equal to round towards +/- infinity.

pub fn overflowing_div_euclid<T>(self, other: T) -> (u256, bool)where
T: Into<u256>,

Calculates the quotient of Euclidean division self.div_euclid(rhs).

Returns a tuple of the divisor along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then self is returned.

pub fn wrapping_div_euclid<T>(self, other: T) -> u256where
T: Into<u256>,

Wrapping Euclidean division. Computes self.div_euclid(rhs), wrapping around at the boundary of the type.

Wrapping will only occur in MIN / -1 on a signed type (where MIN is the negative minimal value for the type). This is equivalent to -MIN, a positive value that is too large to represent in the type. In this case, this method returns MIN itself.

pub fn checked_div_euclid<T>(self, other: T) -> Option<u256>where
T: Into<u256>,

Checked Euclidean division. Computes self.div_euclid(rhs), returning None if rhs == 0 or the division results in overflow.

pub fn rem_euclid<T>(self, other: T) -> u256where
T: Into<u256>,

Calculates the least nonnegative remainder of self (mod rhs).

This is done as if by the Euclidean division algorithm – given r = self.rem_euclid(rhs), self = rhs * self.div_euclid(rhs) + r, and 0 <= r < abs(rhs).

pub fn overflowing_rem_euclid<T>(self, other: T) -> (u256, bool)where
T: Into<u256>,

Overflowing Euclidean remainder. Calculates self.rem_euclid(rhs).

Returns a tuple of the remainder after dividing along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then 0 is returned.

pub fn wrapping_rem_euclid<T>(self, other: T) -> u256where
T: Into<u256>,

Wrapping Euclidean remainder. Computes self.rem_euclid(rhs), wrapping around at the boundary of the type.

Wrapping will only occur in MIN % -1 on a signed type (where MIN is the negative minimal value for the type). In this case, this method returns 0.

pub fn checked_rem_euclid<T>(self, other: T) -> Option<u256>where
T: Into<u256>,

Checked Euclidean remainder. Computes self.rem_euclid(rhs), returning None if rhs == 0 or the division results in overflow.

pub fn checked_shl(self, rhs: u32) -> Option<u256>

Checked shift left. Computes self << rhs, returning None if rhs is larger than or equal to the number of bits in self.

pub fn checked_shr(self, rhs: u32) -> Option<u256>

Checked shift right. Computes self >> rhs, returning None if rhs is larger than or equal to the number of bits in self.

pub fn wrapping_neg(self) -> u256

Wrapping (modular) negation. Computes -self, wrapping around at the boundary of the type. Since unsigned types do not have negative equivalents all applications of this function will wrap (except for -0). For values smaller than the corresponding signed type’s maximum the result is the same as casting the corresponding signed value. Any larger values are equivalent to MAX + 1 - (val - MAX - 1) where MAX is the corresponding signed type’s maximum.

impl u256

pub const MIN: u256 = _

Minimum value

pub const MAX: u256 = _

Maximum value

pub fn is_positive(&self) -> bool

pub fn is_negative(&self) -> bool

pub fn bits_required(&self) -> usize

Return the least number of bits needed to represent the number

pub fn overflowing_mul<T>(self, other: T) -> (u256, bool)where
T: Into<u256>,

Calculates self * rhs

Returns a tuple of the multiplication along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.